+++ extensions/vercast/CHANGELOG.md
@@ -3,7 +3,8 @@
 ## [2.9] - {PR_MERGE_DATE}
 
 - Update items in `Search Components`
-<!-- - Fix extension stuck loading when Token Length is invalid -->
+- Fix extension stuck loading when Token Length is invalid
+- Fix extension would crash when Token invalid
 
 # [2.8] - 2025-02-21